[rename] makeplan -> create-schedule
