+++ src/create-schedule/components/BasicInfo.tsx
