+++ src/create-schedule/components/PlanContent.tsx
