+++ src/create-schedule/components/RemainContent.tsx
