+++ src/create-schedule/components/SideBarIntro.tsx
