+++ src/shared/recoil/index.ts
@@ -1,2 +1,2 @@
 export * from "./modal";
-export * from "./makeplan";
+export * from "./createSchedule";
